+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+ * Include `python_api.h` for TensorFlow ([issue #585](https://github.com/bytedeco/javacpp-presets/issues/585))
+ * Add presets for ONNX 1.2.2 ([pull #547](https://github.com/bytedeco/javacpp-presets/pull/547))
+ * Define `std::vector<tensorflow::OpDef>` type to `OpDefVector` for TensorFlow
+ * Link HDF5 with zlib on Windows also ([issue deeplearning4j/deeplearning4j#6017](https://github.com/deeplearning4j/deeplearning4j/issues/6017))
+ * Enable MKL-DNN for TensorFlow
+ * Upgrade presets for TensorFlow 1.10.0-rc1
  * Fix loading issue with `opencv_cudaobjdetect` and `opencv_cudaoptflow` on Windows ([issue #592](https://github.com/bytedeco/javacpp-presets/issues/592))
 
 ### July 17, 2018 version 1.4.2